replace vecan-mk2 with mk2-dbus
The mk2 interfacing code is called vecan-mk2, since that project is originally
used for this code. Its name is confusing though since it has nothing to do with
vecan in the venus build, so it is renamed to mk2-dbus.

note: there are several other problems with the start / stop script, /service/vecan
doesn't exists. Killing all vedirect services will simply respawn them etc. I am
ignoring that for now, since time is better spend in how to support demo mode in
a better way.