handle ItemsChanged signals
The previous PropertiesChanged on root implementation has been
replaced with an ItemsChanged signal that does the same job.
This time, instead of recording the message as is, we
instead unroll it. This will be compensated for during replay
in giving the user the option to replay the recording in packed
ItemsChanged signals or individual PropertiesChanged signals.

https://github.com/victronenergy/venus/issues/789