play.py: fix stupid bug in service reset code
when resetting service state to startup values (at the end of the
loop), a bunch of empty dicts were sent, one for every path, in
really quick succession.  Also, values were included for services
other than the one being reset, because this idiot initialised the
dictionary outside the proper loop.